+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/BangunLingkaran.kt
@@ -1,5 +1,5 @@
 package com.mopro.rumusdasarmatermatika.model
 
 data class BangunLingkaran(
-    val jari: Long
+    val jari: Float
 )
