+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/BangunPersegi.kt
@@ -1,5 +1,5 @@
 package com.mopro.rumusdasarmatermatika.model
 
 data class BangunPersegi(
-    val sisi: Long
+    val sisi: Float
 )
