+++ app/src/main/java/com/mopro/rumusdasarmatermatika/model/BangunPersegiPanjang.kt
@@ -1,6 +1,6 @@
 package com.mopro.rumusdasarmatermatika.model
 
 data class BangunPersegiPanjang(
-    val panjang: Long,
-    val lebar: Long
+    val panjang: Float,
+    val lebar: Float
 )
